requirements.txt
pytest==7.3.1
selene==2.0.0b17
allure-pytest
python-dotenv